refactor(converters): scope mfenced, mtext, mo, mi and generic-under-over into folders
Promote each converter that owns auxiliary classes into its own folder, mirroring
the existing mfrac layout: a re-exporting index.ts, the main converter file and one
file per auxiliary class (Vector/Matrix/Separators, TextCommand, Operator, Character,
UnderOverSetter).
