+++ src/data/usecases/mathml-to-latex-convertion/converters/generic-under-over/index.ts
@@ -1,0 +1,1 @@
+export { GenericUnderOver } from './generic-under-over';
