+++ src/data/usecases/mathml-to-latex-convertion/converters/generic-under-over/mover.integration.test.ts
@@ -1,5 +1,5 @@
-import { MathMLToLaTeX } from '../../../..';
-import { InvalidNumberOfChildrenError } from '../../../errors/invalid-number-of-children';
+import { MathMLToLaTeX } from '../../../../..';
+import { InvalidNumberOfChildrenError } from '../../../../errors/invalid-number-of-children';
 
 describe('mover (integration)', () => {
   it.each([
