+++ src/data/usecases/mathml-to-latex-convertion/converters/generic-under-over/munder.integration.test.ts
@@ -1,4 +1,4 @@
-import { MathMLToLaTeX } from '../../../..';
+import { MathMLToLaTeX } from '../../../../..';
 
 describe('munder (integration)', () => {
   it.each([
