+++ src/data/usecases/mathml-to-latex-convertion/converters/mfenced/index.ts
@@ -1,0 +1,1 @@
+export { MFenced } from './mfenced';
