+++ src/data/usecases/mathml-to-latex-convertion/converters/mfenced/mfenced.integration.test.ts
@@ -1,4 +1,4 @@
-import { MathMLToLaTeX } from '../../../..';
+import { MathMLToLaTeX } from '../../../../..';
 
 describe('mfenced (integration)', () => {
   it.each([
