+++ src/data/usecases/mathml-to-latex-convertion/converters/mi/index.ts
@@ -1,0 +1,1 @@
+export { MI } from './mi';
