+++ src/data/usecases/mathml-to-latex-convertion/converters/mi/mi.integration.test.ts
@@ -1,4 +1,4 @@
-import { MathMLToLaTeX } from '../../../..';
+import { MathMLToLaTeX } from '../../../../..';
 
 describe('mi (integration)', () => {
   it.each([
