+++ src/data/usecases/mathml-to-latex-convertion/converters/mi/mi.test.ts
@@ -1,5 +1,5 @@
 import { MI } from './mi';
-import { MathMLElement } from '../../../protocols/mathml-element';
+import { MathMLElement } from '../../../../protocols/mathml-element';
 
 const mi = (value: string, attributes: Record<string, string> = {}): MathMLElement => ({
   name: 'mi',
