+++ src/data/usecases/mathml-to-latex-convertion/converters/mi/mi.ts
@@ -1,9 +1,9 @@
-import { ToLaTeXConverter } from '../../../../domain/usecases/to-latex-converter';
-import { MathMLElement } from '../../../protocols/mathml-element';
-import { normalizeWhiteSpaces, ownLookup } from '../../../helpers';
-import { allMathSymbolsByChar, allMathSymbolsByGlyph, mathNumberByGlyph } from '../../../../syntax';
+import { ToLaTeXConverter } from '../../../../../domain/usecases/to-latex-converter';
+import { MathMLElement } from '../../../../protocols/mathml-element';
+import { normalizeWhiteSpaces } from '../../../../helpers';
 import { UTF8ToLtXConverter } from 'data/protocols';
-import { HashUTF8ToLtXConverter } from '../../../../syntax/utf8-converter';
+import { HashUTF8ToLtXConverter } from '../../../../../syntax/utf8-converter';
+import { Character } from './character';
 
 /**
  * Converts a MathML `<mi>` (identifier) element into LaTeX.
@@ -77,39 +77,5 @@ export class MI implements ToLaTeXConverter {
       default:
         return value;
     }
-  }
-}
-
-/** Resolves a single identifier value to its LaTeX command via symbol lookup tables, falling back to UTF-8 conversion. */
-class Character {
-  private _value: string;
-
-  constructor(value: string) {
-    this._value = value;
-  }
-
-  static apply(value: string): string {
-    return new Character(value)._apply();
-  }
-
-  private _apply(): string {
-    return (
-      this._findByCharacter() ||
-      this._findByGlyph() ||
-      this._findByNumber() ||
-      new HashUTF8ToLtXConverter().convert(this._value)
-    );
-  }
-
-  private _findByCharacter(): string | undefined {
-    return ownLookup(allMathSymbolsByChar, this._value);
-  }
-
-  private _findByGlyph(): string | undefined {
-    return ownLookup(allMathSymbolsByGlyph, this._value);
-  }
-
-  private _findByNumber(): string | undefined {
-    return ownLookup(mathNumberByGlyph, this._value);
   }
 }
