+++ src/data/usecases/mathml-to-latex-convertion/converters/mo/mo.integration.test.ts
@@ -1,4 +1,4 @@
-import { MathMLToLaTeX } from '../../../..';
+import { MathMLToLaTeX } from '../../../../..';
 
 describe('mo (integration)', () => {
   it.each([
