+++ src/data/usecases/mathml-to-latex-convertion/converters/mtext/index.ts
@@ -1,0 +1,1 @@
+export { MText } from './mtext';
