+++ src/data/usecases/mathml-to-latex-convertion/converters/mtext/mtext.integration.test.ts
@@ -1,4 +1,4 @@
-import { MathMLToLaTeX } from '../../../..';
+import { MathMLToLaTeX } from '../../../../..';
 
 describe('mtext (integration)', () => {
   it.each([
